Progress toward removing SageMath and converting as much code to Python as possible.
CreateFiles.py will create one of three types of files: Exam, Key, or Feedback.
    Exam is the traditional multiple-choice exam students see.
    Key is the generic key that will be posted for general use.
    Feedback is the specific key for each student.
This also houses the functions to close a file once it is finished.

manipulateFiles.py will copy keys from the generation folder to the CompleteExam folder.
I need to include the same for Exam pdfs and Feedback pdfs.
This also clears all old versions of exams.

pythonImports.py is a list of python imports that should be included in every question's code.

retrieveQuestionListsFromDatabase.py lists all questions stored in the database. This will be used for Zenity to display the question options.

intervalMaskingMethod.py is Russell's original masking method. This will need to be revised to dynamically determine the interval range and avoid infinite decents because options are too close to one another. It should also provide different intervals when there are copies in the optionList.

printQuestions.py is the new method for printing questions to the Exam, Key, and Feedback. The list of required values from the database is provided in the file.

saveQuestionListsToDatabase.py saves questions by topic. This will be used in Zenity to add questions to the database to dynamically list available questions.

TO DO:
1. Convert generateExams.sh to python functions.
2. Convert functionsForZenityAutoDIGshFile.sh to python functions, where appropriate.
3. Convert as much zenityAutoDIG.sh to python as possible.
4. Add pythonImports.py, intervalMaskingMethod.py, and commonlyUsedFunctions.py to beginning of each question code.
5. Ensure each question code contains necessary variables to be saved to the question database.
6. Debug any issues that arise.